clean QuickAdd page
